Do not close std* file descriptors twice
When the process is started, we close the file descriptors, but do not
set them as invalid. If the process is never started (session is
terminated in waiting state), we close the fds in desctructor. This
destructor closing lead to double close when the session started the
process. The result was, that we likely closed some other file
descriptors.